+++ jslint.mjs
@@ -120,39 +120,40 @@
     message, meta, min, mkdir, modeCoverageIgnoreFile, modeIndex, mode_cli,
     mode_conditional, mode_json, mode_module, mode_noop, mode_property,
     mode_shebang, mode_stop, module, moduleFsInit, moduleName, module_list,
-    name, names, node, noop, now, nr, nud_prefix, objectDeepCopyWithKeysSorted,
-    ok, on, open, opening, option, option_dict, order, package_name, padEnd,
-    padStart, parameters, parent, parentIi, parse, pathname, pathnameList,
-    platform, pop, processArgv, process_argv, process_env, process_exit,
-    promises, property, property_dict, push, quote, ranges, readFile, readdir,
-    readonly, recursive, reduce, repeat, replace, resolve, result, reverse,
-    role, round, scriptId, search, set, shebang, shift, signature, single,
-    slice, some, sort, source, spawn, splice, split, stack, stack_trace, start,
-    startOffset, startsWith, statement, statement_prv, stdio, stop, stop_at,
-    stringify, switch, syntax_dict, tenure, test, test_cause,
-    test_internal_error, this, thru, toString, token, token_global, token_list,
-    token_nxt, token_tree, tokens, trace, tree, trim, trimEnd, trimRight, try,
-    type, unlink, unordered, unshift, url, used, v8CoverageListMerge,
-    v8CoverageReportCreate, value, variable, version, versions, warn, warn_at,
-    warning, warning_list, warnings, white, wrapped, writeFile
+    name, names, node, nomen, noop, now, nr, nud_prefix,
+    objectDeepCopyWithKeysSorted, ok, on, open, opening, option, option_dict,
+    order, package_name, padEnd, padStart, parameters, parent, parentIi, parse,
+    pathname, pathnameList, platform, pop, processArgv, process_argv,
+    process_env, process_exit, promises, property, property_dict, push, quote,
+    ranges, readFile, readdir, readonly, recursive, reduce, repeat, replace,
+    resolve, result, reverse, role, round, scriptId, search, set, shebang,
+    shift, signature, single, slice, some, sort, source, spawn, splice, split,
+    stack, stack_trace, start, startOffset, startsWith, statement,
+    statement_prv, stdio, stop, stop_at, stringify, subscript, switch,
+    syntax_dict, tenure, test, test_cause, test_internal_error, this, thru,
+    toString, token, token_global, token_list, token_nxt, token_tree, tokens,
+    trace, tree, trim, trimEnd, trimRight, try, type, unlink, unordered,
+    unshift, url, used, v8CoverageListMerge, v8CoverageReportCreate, value,
+    variable, version, versions, warn, warn_at, warning, warning_list, warnings,
+    white, wrapped, writeFile
 */
 
 // init debugInline
 let debugInline = (function () {
-    let consoleError = function () {
+    let __consoleError = function () {
         return;
     };
     function debug(...argv) {
 
 // This function will print <argv> to stderr and then return <argv>[0].
 
-        consoleError("\n\ndebugInline");
-        consoleError(...argv);
-        consoleError("\n");
+        __consoleError("\n\ndebugInline");
+        __consoleError(...argv);
+        __consoleError("\n");
         return argv[0];
     }
     debug(); // Coverage-hack.
-    consoleError = console.error;
+    __consoleError = console.error;
     return debug;
 }());
 let jslint_charset_ascii = (
@@ -3285,9 +3286,10 @@ function jslint_phase2_lex(state) {
         case "getset":          // Allow get() and set().
         case "indent2":         // Use 2-space indent.
         case "long":            // Allow long lines.
-        case "name":            // Allow weird property names.
         case "node":            // Assume Node.js environment.
+        case "nomen":           // Allow weird property names.
         case "single":          // Allow single-quote strings.
+        case "subscript":       // Allow identifiers in subscript-notation.
         case "test_cause":      // Test jslint's causes.
         case "test_internal_error":     // Test jslint's internal-error
                                         // ... handling-ability.
@@ -3300,6 +3302,12 @@ function jslint_phase2_lex(state) {
         case "white":           // Allow messy whitespace.
             option_dict[key] = val;
             break;
+        // alias for eval
+        case "evil":
+            return option_set_item("eval", val);
+        // alias for nomen
+        case "name":
+            return option_set_item("nomen", val);
         default:
             return false;
         }
@@ -3388,6 +3396,7 @@ console.log(JSON.stringify(Object.keys(window).sort(), undefined, 4));
                 "location",
                 // "name",
                 "navigator",
+                "postMessage",
                 // "screen",
                 "sessionStorage",
                 // "setInterval",
@@ -4545,7 +4554,12 @@ function jslint_phase3_parse(state) {
         let the_subscript = parse_expression(0);
         if (the_subscript.id === "(string)" || the_subscript.id === "`") {
             name = survey(the_subscript);
-            if (jslint_rgx_identifier.test(name)) {
+
+// PR-xxx - directive
+// Add new directive "subscript" allowing JSLint to be used with scripts
+// targeting Google Closure Compiler.
+
+            if (!option_dict.subscript && jslint_rgx_identifier.test(name)) {
 
 // test_cause:
 // ["aa[`aa`]", "infix_lbracket", "subscript_a", "aa", 4]
@@ -6700,12 +6714,17 @@ function jslint_phase3_parse(state) {
 // Restore previous catch-scope after catch-block.
 
             catchage = catch_stack.pop();
-        } else {
 
-// test_cause:
-// ["try{}finally{break;}", "stmt_try", "expected_a_before_b", "finally", 6]
+// PR-xxx - warning
+// Relax warning about missing `catch` in `try...finally` statement.
+//
+//         } else {
+//
+// // test_cause:
+// // ["try{}finally{break;}", "stmt_try", "expected_a_before_b", "finally", 6]
+//
+//             warn("expected_a_before_b", token_nxt, "catch", artifact());
 
-            warn("expected_a_before_b", token_nxt, "catch", artifact());
         }
         if (token_nxt.id === "finally") {
             functionage.finally += 1;
@@ -7054,7 +7073,7 @@ function jslint_phase3_parse(state) {
                     warn("unregistered_property_a", name);
                 }
             } else if (
-                !option_dict.name
+                !option_dict.nomen
                 && name.identifier
                 && jslint_rgx_weird_property.test(id)
             ) {