+++ jslint_ci.sh
@@ -1713,7 +1713,11 @@ function objectDeepCopyWithKeysSorted(obj) {
             flags
         }) {
             result0 = result;
-            result = result.replace(new RegExp(aa, flags), bb);
+            result = result.replace(new RegExp(aa, flags), bb.replace((
+                /\*\\\\\//g
+            ), "*/").replace((
+                /\/\\\\\*/g
+            ), "/*"));
             if (result0 === result) {
                 throw new Error(
                     "shRawLibFetch - cannot find-and-replace snippet "