+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 # Changelog
 
 # Todo
-- jslint - allow alias `nomen` for jslint-directive `name`
 - wrapper - add vscode-command to suppress minor warnings on given line
 - doc - document jslint directives and supported/unsupported es6+ features
 - cli - remove cli-option `--mode-vim-plugin`
@@ -16,6 +15,10 @@
 - node - after node-v14 is deprecated, remove shell-code `export "NODE_OPTIONS=--unhandled-rejections=strict"`.
 
 # v2022.6.1-beta
+- directive - add new directive `subscript` allowing JSLint to be used with scripts targeting Google Closure Compiler
+- warning - relax warning about missing `catch` in `try...finally` statement
+- jslint - allow aliases `evil, nomen` for jslint-directives `eval, name`, respectively
+- bugfix - fix broken codemirror example
 - bugfix - fix jslint not-recognizing option-chaining when comparing operands of binary operator
 - allow array-literals to directly call [...].flat() and [...].flatMap()